Time: 209 ms (33.96%), Space: 51.3 MB (21.31%) - LeetHub

+++ 1653-minimum-deletions-to-make-string-balanced/1653-minimum-deletions-to-make-string-balanced.cpp
@@ -23,9 +23,9 @@ public:
             }
         }
         
-        long long ans = s.length();
+        int ans = s.length();
         
-        long long tmp = s.length()-postfixcnt_b[0];
+        int tmp = s.length()-postfixcnt_b[0];
            ans = min(ans,tmp);
         // cout<<"postfixcnt_b[0] ="<<postfixcnt_b[0]<<endl;
         tmp = s.length()-prefixcnt_a[s.length()-1];
